backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
qrcode[pil]==7.4.2
email-validator==2.1.1
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9

pydantic==2.6.4
pydantic-settings==2.2.1

pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
scipy==1.13.0
scikit-image==0.23.2
pytesseract==0.3.10
piexif==1.1.3

reportlab==4.1.0

httpx==0.27.0
python-dotenv==1.0.1
aiofiles==23.2.1
websockets==12.0
loguru==0.7.2

gunicorn==21.2.0
pyzbar==0.1.9
huggingface-hub==0.21.4
python-dotenv==1.0.1